move TestDeleteEventDrop_Issue18089 to reproduce_18089_test.go
